Refactor: move getFontList() to UI
