+++ src/background/utils/extension-api.ts
@@ -97,27 +97,6 @@ export const subscribeToOuterSettingsChange = (callback: () => void) => {
     });
 };
 
-export async function getFontList() {
-    return new Promise<string[]>((resolve) => {
-        if (!chrome.fontSettings) {
-            // Todo: Remove it as soon as Firefox and Edge get support.
-            resolve([
-                'serif',
-                'sans-serif',
-                'monospace',
-                'cursive',
-                'fantasy',
-                'system-ui'
-            ]);
-            return;
-        }
-        chrome.fontSettings.getFontList((list) => {
-            const fonts = list.map((f) => f.fontId);
-            resolve(fonts);
-        });
-    });
-}
-
 export async function getCommands() {
     return new Promise<chrome.commands.Command[]>((resolve) => {
         if (!chrome.commands) {
