+++ src/ui/popup/components/more-settings/index.tsx
@@ -9,7 +9,7 @@ import {getLocalMessage} from '../../../../utils/locales';
 import type {ExtWrapper, FilterConfig, TabInfo} from '../../../../definitions';
 import {isFirefox} from '../../../../utils/platform';
 
-export default function MoreSettings({data, actions, tab}: ExtWrapper & {tab: TabInfo}) {
+export default function MoreSettings({data, actions, tab, fonts}: ExtWrapper & {tab: TabInfo} & {fonts: string[]}) {
     const custom = data.settings.customThemes.find(({url}) => isURLInList(tab.url, url));
     const filterConfig = custom ? custom.theme : data.settings.theme;
 
@@ -25,7 +25,7 @@ export default function MoreSettings({data, actions, tab}: ExtWrapper & {tab: Ta
     return (
         <section class="more-settings">
             <div class="more-settings__section">
-                <FontSettings config={filterConfig} fonts={data.fonts} onChange={setConfig} />
+                <FontSettings config={filterConfig} fonts={fonts} onChange={setConfig} />
             </div>
             <div class="more-settings__section">
                 {isFirefox ? null : <p class="more-settings__description">
