+++ src/ui/popup/index.tsx
@@ -6,8 +6,9 @@ import {popupHasBuiltInHorizontalBorders, popupHasBuiltInBorders, fixNotClosingP
 import type {ExtensionData, ExtensionActions, TabInfo} from '../../definitions';
 import {isMobile, isFirefox} from '../../utils/platform';
 import {MessageType} from '../../utils/message';
+import {getFontList} from '../utils';
 
-function renderBody(data: ExtensionData, tab: TabInfo, actions: ExtensionActions) {
+function renderBody(data: ExtensionData, tab: TabInfo, fonts: string[], actions: ExtensionActions) {
     if (data.settings.previewNewDesign) {
         if (!document.documentElement.classList.contains('preview')) {
             document.documentElement.classList.add('preview');
@@ -17,7 +18,7 @@ function renderBody(data: ExtensionData, tab: TabInfo, actions: ExtensionActions
     }
 
     sync(document.body, (
-        <Body data={data} tab={tab} actions={actions} />
+        <Body data={data} tab={tab} actions={actions} fonts={fonts}/>
     ));
 }
 
@@ -25,12 +26,13 @@ async function start() {
     const connector = connect();
     window.addEventListener('unload', () => connector.disconnect());
 
-    const [data, tab] = await Promise.all([
+    const [data, tab, fonts] = await Promise.all([
         connector.getData(),
         connector.getActiveTabInfo(),
+        getFontList()
     ]);
-    renderBody(data, tab, connector);
-    connector.subscribeToChanges((data) => renderBody(data, tab, connector));
+    renderBody(data, tab, fonts, connector);
+    connector.subscribeToChanges((data) => renderBody(data, tab, fonts, connector));
 }
 
 addEventListener('load', start);
